refactor(lib): extract text/width, short-id, relative-time, origin-machine out of session/
Move 1 of the core-module refactor. Session files stay as export * shims.
Sessions impact budget_sec 180→240 (run 32392267349 measured 198s); pin
ci-scope.test.ts. Replayed onto current main so CI does not select the
279-file two-dot set from PR #2796's merge commit (run 32433778106, 550s).
