+++ apps/cli/ci/test-ownership.yaml
@@ -98,13 +98,17 @@ groups:
   # typecheck + bench. It could never pass 85s, so no new `agents sessions <verb>`
   # was mergeable; cli-full's 1200s would have been the opposite overcorrection.
   # Raise this only against a measured run, and lower it when the suite is split.
+  # 240s, not 180s. Extracting width/short-id/relative-time out of session/
+  # (PR #2796, run 32392267349) still touches session/* re-export shims, so
+  # the sessions group is selected; impact was 198s of a passing vitest run
+  # and failed the 180s ceiling. Same pattern as daemon 240 / hooks 150.
   - id: sessions
     when:
       - apps/cli/src/lib/session/**
       - apps/cli/src/commands/sessions.ts
       - apps/cli/bench/sessions-active-perf.ts
     checks: [typecheck, sessions-bench]
-    budget_sec: 180
+    budget_sec: 240
 
   # 240s, not the default 85s. Moving runner.ts into daemon/ (PR #2803, run
   # 32395701692) selected companion + importer tests including routines.test.ts
