+++ apps/cli/src/lib/session/remote.ts
@@ -32,8 +32,8 @@ import { SSH_OPTS, controlOpts, assertValidSshTarget } from '../ssh-exec.js';
 import { remoteShellFor, buildWindowsAgentsCommand } from '../hosts/remote-cmd.js';
 import { resolveRemoteOsSync } from '../hosts/remote-os.js';
 import { NO_FANOUT_ENV } from './remote-active.js';
-import { formatRelativeTime } from './relative-time.js';
-import { terminalWidth } from './width.js';
+import { formatRelativeTime } from '../text/relative-time.js';
+import { terminalWidth } from '../text/width.js';
 
 /**
  * POSIX single-quote a string for safe interpolation into a remote shell command.
