+++ apps/cli/src/lib/text/short-id.test.ts
