+++ apps/cli/src/lib/text/width.test.ts
